% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCInfPred.R
\name{AUCInfPred}
\alias{AUCInfPred}
\title{Calculate Area Under Curve or Moment Curve To "Infinity" (Predicted)}
\usage{
AUCInfPred(conc, time, lamznpt = NULL, lambdaZStats = NULL,
  calculation = c("standard", "moment"), minpoints = 3, addt0 = FALSE,
  inter = "Linear")
}
\arguments{
\item{conc}{A numeric vector of concentration values}

\item{time}{A numeric vector of time values, parallel to \code{conc} and of the same length.  These should be sorted
in ascending order.}

\item{lamznpt}{Number of points to use for the lambdaz calculation, counted from the end of the concentration/time vectors.}

\item{lambdaZStats}{if not \code{NULL}, a list with 9 elements (see details). Must be provided if lamznpt is not.}

\item{calculation}{Must be either the string "standard" or "moment".  If the former, calculates the standard area under
the curve.  For the latter, it will calculate the area under the moment curve}

\item{minpoints}{Minimum number of points to use for the Lambda-z calculation(s), 3 by default. Single positive integer.}

\item{addt0}{Single logical value declaring whether T0 should be added if missing (default \code{TRUE}).}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default) or \code{"Linear Log"}}
}
\value{
A numeric vector with a single element holding the area under the concentration/time curve until "infinity".
}
\description{
Area Under the concentration time Curve from time = 0 to time = Inf using extrapolation from Predicted Clast
}
\details{
AUCInfPred calculates the area under a time/concentration (or first moment) curve from the first point up to  
"infinity" using the linear trapezium rule.  This is performed by calculating the area under the concentration-time 
curve until the last measurable concentration (see \code{\link{AUCLast}}), and then adding an extrapolated 
AUC which is the area bounded by the exponential decay curve with rate constant -lambdaz from observed Clast.  
Analogous calculations for AUMC are used if appropriate. \cr
There are two methods for calculating AUC and AUMC.
LAMZstats can be calculated within the function. In this case, lamznpt should be provided.
lamznpt must be a single integer greater than minPonts (which should be at least 2) and less than or equal to the length of \code{time} and \code{conc}.
lamznpt is NULL by default. Must be provided if lambdaZStats is not.
The lambdaz statistics may be provided directly to this function; it should have length 1 numeric elements:
\enumerate{
     \item LAMZ
     \item intercept
     \item R2
     \item R2ADJ
     \item CORRXY
     \item LAMZHL
     \item LAMZLL
     \item LAMZUL
     \item lamznpt
}
Use \code{unlist} to pass \code{lambdaZStatistics} output to this function.
}
\note{
The following error checks / processing will be performed:
 \enumerate{
     \item all error checks for \code{AUCLast} apply
     \item If any relevant lambda z calculations are NA, NA will be returned
     \item An exception will be generated if lamznpt is not a single integer numeric between 2 and  \code{length(conc)} (inclusive)
 }
The algorithm for calculating the extrapolated AUC is as follows:
\deqn{ClastExtrap = \exp(-\lambda_z * t_{upper}) * c_0} 
\deqn{AUCPredInf = AUCLast + ClastExtrap / \lambda_z}   
For the extrapolated AUMC, the calculation is
\deqn{AUMCPredInf = AUMCLast + ClastExtrap * tlast / \lambda_z + ClastExtrap / \lambda_z^2}
Here \eqn{c_0} is the exponential of the intercept term from the lambda z regression, and \eqn{t_{upper}} is the
last time point used in that regression
}
\examples{
     Theoph1 <- subset(Theoph, Subject == 1)
     AUCInfPred(Theoph1$conc, Theoph1$time, lamznpt = 4)
}
\seealso{
\code{\link{lambdaZStatistics}}, \code{\link{AUCLast}}
}
\author{
Mango Solutions
}
\keyword{math}
