% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCPartial.R
\name{AUCPartial}
\alias{AUCPartial}
\title{Partial Area Under Concentration Time Curve}
\usage{
AUCPartial(Conc, Time, endTime, numPoints = NULL, lambdaZStats = NULL,
  usePoints = NULL, excPoints = FALSE, minPoints = 3, addT0 = FALSE,
  inter = "Linear")
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{endTime}{: last time at which area should be calculated}

\item{numPoints}{: Number of points to use for lambdaz calculation.}

\item{lambdaZStats}{if not \code{NULL}, a list with 9 elements (see details). Must be provided if numPoints is not.}

\item{usePoints}{If NULL (default) automatically select, else, logical vector of points to use (\code{TRUE}) for calculation of terminal phase.}

\item{excPoints}{Logical vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}

\item{minPoints}{single numeric, the fewest points for which a linear model fit should be attempted (default 3)}

\item{addT0}{single logical, should T=0, C=0 be added to the input data if T0 is missing? (default \code{FALSE})}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}
}
\value{
Single numeric with partial area under curve
}
\description{
Area Under the Concentration Time Curve from T = 0 to T = endTime
}
\details{
Calculates the area under a time-concentration curve from the first time up until the \code{endTime}.  
\code{endTime} need not be one of the elements of \code{Time}, but it should not lie before the minimum time.
If the endTime does not coincide with an existing time element and is less than tlast, the following 
interpolation formula will be used to calculated a new concentration: 
\deqn{c_{inter} = c_1 +  \left| \frac{ t^{end} - t_1}{t_2 - t_1} \right|  (c_2 - c_1)}.  
If endTime is greater than tlast the extrapolated area is calculated using \code{getTerminalAUC} 
which calculates the area under the terminal phase exponential from TLast to endTime.
\code{numPoints} is calculated by \code{\link{getPartialAUC}} .  
This calculation is the integral of the Terminal phase exponential decay function: 
\deqn{\int intercept \ast exp\left ( -\lambda _{z} \ast T\right ) dT = 
    -intercept / \left ( \lambda _{z} \ast exp\left ( \lambda _{z} \ast T\right )\right)}
The partial area under the terminal concentration time curve can be calculated 
by subtracting modeled AUC0Tlast from modeled AUC0endTime.
\code{numPoints} may be NA provided endTime < Tlast.
}
\seealso{
\code{\link{lambdaZStastics}} for more about the numPoints parameter
}
\author{
Mango Business Solutions
}
\keyword{math}
