% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCPP-class.R
\docType{class}
\name{CDISCPP-class}
\alias{CDISCPP-class}
\alias{CDISCPP}
\title{An S4 class to represent CDISCPP data.}
\description{
A data frame-like object with 38 named columns.
}
\examples{
x <- new("CDISCPP")
print(x)
}
