% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaComplete.R
\name{chooseNumPointsAction}
\alias{chooseNumPointsAction}
\title{Choose how to handle lamznpt}
\usage{
chooseNumPointsAction(conc = conc, time = time, lamznpt = NA_real_,
  usepoints = NULL, excpoints = FALSE)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}

\item{usepoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to use for calculation of terminal phase.
Used rows are flagged by usepoints as \code{TRUE}.}

\item{excpoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to exclude from automatic calculation of terminal phase.
Excluded rows are flagged by excpoints as \code{TRUE}.}
}
\value{
length 2 list with elements ACTION, a single character value,
and MINROWSFORLAMBDAZ, a single numeric with value 3
}
\description{
Decide whether to use or automatically calculate lamznpt
}
\author{
Mango Solutions
}
