% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasttlast.R
\name{ClastTlast}
\alias{ClastTlast}
\title{Calculate Last Measurable concentration}
\usage{
ClastTlast(conc, time)
}
\arguments{
\item{conc}{numeric vector of concentrations}

\item{time}{numeric vector of time points (should be of equal length to conc, and sorted).}
}
\value{
list with numeric scalar components \code{clast}, \code{tlast} and \code{index}
}
\description{
Finds the last measurable (non-zero) concentration value and the last time point
with that concentration.
}
\note{
The following error checks / processing will be performed:
 \enumerate{
     \item If \code{conc} or \code{time} are of length 0, a single NA is returned
     \item \code{conc} and \code{time} should be equal length numeric vectors, otherwise an exception will be generated
     \item It is assumed that these times and concentrations come from a single curve 
     \item time should be sorted
     \item If the sum of \code{conc} (after omitting missing values) is 0, returned values are NA
 }
}
\examples{
Theoph1 <- subset(  Theoph, Subject == 1)
ClastTlast( conc = Theoph1$conc, time = Theoph1$time )
}
\keyword{math}
\keyword{utils}
