% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearance.R
\name{CLObs}
\alias{CLObs}
\title{Apparent Clearance (Observed)}
\usage{
CLObs(conc, time, dose, lamznpt = NULL)
}
\arguments{
\item{conc}{A numeric vector of concentration values}

\item{time}{A numeric vector of time values, parallel to \code{conc} and of the same length.  These should be sorted
in ascending order.}

\item{dose}{A single numeric}

\item{lamznpt}{Number of points to use for the lambda z calculation, counted from the last measurable concentration.
Must be a single integer greater than 1 and less than or equal to the length of \code{time} and \code{conc}.
Default is NULL, lamznpt is then calculated by \code{selectPoints}.}
}
\value{
apparent clearance (single numeric)
}
\description{
Calculates apparent clearance (observed), which is equal to administered dose divided by AUC0Inf_Obs.
}
\note{
The following additional input processing/checks are performed:
\enumerate{
     \item All checks for \code{AUCInfObs} apply
     \item \code{dose} must be a numeric vector of length 1, otherwise an exception will be generated.  If
         \code{dose} is \code{NA}, 0 or less, \code{NA} will be returned 
}
}
\examples{
Theoph1 <- subset(  Theoph, Subject == 1)
CLObs( conc = Theoph1$conc, time = Theoph1$time, dose = Theoph1$Dose[1])
}
\author{
Mango Solutions
}
\keyword{math}
