% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaxtmax.R
\name{CmaxTmax}
\alias{CmaxTmax}
\title{Calculate Maximum Concentration and time of maximum concentration}
\usage{
CmaxTmax(Conc, Time)
}
\arguments{
\item{Conc}{numeric vector of concentrations}

\item{Time}{numeric vector of time points (parallel to \code{Conc})}
}
\value{
list with numeric scalar (length 1 vector) components "cmax", "tmax" and "index". See notes for further processing.
}
\description{
Calculate Maximum Concentration and time of maximum concentration
}
\details{
Finds the maximum concentration value and the first time point 
(in the sense of corresponding to the minimal index)
with that concentration.
}
\note{
The following error checks / processing will be performed:
 \enumerate{
     \item If \code{Conc} or \code{Time} are of length 0, NA is returned for both values
     \item If the sum of \code{Code} (after omitting missing values) is 0, NA is returned for both values
     \item If \code{cmax} is 0, NA is returned for \code{tmax}
     \item \code{cmax} is computed by stripping NA values from \code{Conc}
     \item if \code{Conc} and \code{Time} are not equal length numeric vectors, an exception will be generated by
           \code{\link{checkNumericSameLength}}
 }
}
\examples{
Theoph.2 <- subset(Theoph, Subject == 2)
CmaxTmax( Theoph.2$conc, Theoph.2$Time ) 
}
\keyword{math}
