% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPoints.R
\name{fixedPoints}
\alias{fixedPoints}
\title{Fixed Points for Calculating Terminal Rate Constant}
\usage{
fixedPoints(Conc, Time, numPoints, excPoints = FALSE, minPoints = 3,
  Safe = TRUE)
}
\arguments{
\item{Conc}{Vector of concentrations.}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates.}

\item{excPoints}{Logical vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}

\item{minPoints}{Minimum number of points to use for the Lambdaz calculation(s), 3 by default. Single positive integer.}

\item{Safe}{Single logical value declaring whether to perform redundant data checks (default is TRUE).}

\item{minPoints}{Single numeric declaring minimum acceptable number of rows for linear model.}
}
\value{
List consisting of numPoints, a single numeric value, and result, a length 9 numeric vector.
}
\description{
Use Fixed Number of Trailing Points for Terminal Rate Constant (Lambdaz) Calculation.
}
\details{
Returns input numPoints and results for Terminal Rate Constant (Lambdaz) calculation
fixedPoints calls \code{\link{lambdaZStatistics}} to perform regressions using the specified number of trailing non-zero rows
Rows prior to Cmax may be used.
Exclusion of rows prior to the end of infusion is not currently implemented.
Zero or missing values will error.
}
\examples{
     Theoph1 <- subset(Theoph, Subject == 1)
     fixedPoints(Conc = Theoph1$conc, Time = Theoph1$Time, numPoints = 3)
}
\author{
Mango Solutions
}
\keyword{math}
