% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConcentration.R
\name{getConcentration}
\alias{getConcentration}
\title{Concentration at Time at T = firstTime and T = secondTime}
\usage{
getConcentration(Conc = NULL, Time = NULL, firstTime = NA,
  secondTime = NA, numPoints = NULL, usePoints = NULL, excPoints = NULL,
  Safe = TRUE, inter = "Linear")
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{firstTime}{single numeric value, first time at which Concentration should be calculated}

\item{secondTime}{single numeric value, second time at which Concentration should be calculated}

\item{numPoints}{single numeric value declaring number of points to use}

\item{usePoints}{If \code{NULL} (default) automatically select, else, logical vector of points to use for calculation of terminal phase. Used rows are flagged by usePoints as \code{TRUE}.}

\item{excPoints}{If \code{NULL} (default) automatically select, else, logical vector of points to exclude from automatic calculation of terminal phase. Excluded rows are flagged by excPoints as \code{TRUE}.}

\item{Safe}{single logical value declaring whether error checking should be performed}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}
}
\value{
Data frame
}
\description{
Concentration at Time at T = firstTime and T = secondTime
}
\details{
This function calculates Concentration at Time
Returns a data frame with 1 row containing columns:
 \enumerate{
     \item \code{firstTime}  
     \item \code{secondTime} 
     \item \code{firstConc}
     \item \code{secondConc}
     \item \code{numPoints} 
     \item \code{Error} 
 }
}
\author{
Mango Solutions
}
\keyword{math}
