% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaz.R
\name{getLambdaz}
\alias{getLambdaz}
\title{Perform \code{lm} on logconc-time data

Calculates lambdaz for a data frame consisting of logconc and time.
The \code{lm} function is used to fit a linear regression of the last \code{lamznpt}
concentration against time.  These are counted from (and include)
the last measurable concentration.}
\usage{
getLambdaz(timeconc)
}
\arguments{
\item{timeconc}{Data frame with columns logconc and time used for calculating lambdaz.}
}
\value{
\code{"lm"} object to be used by \code{getLambdaZStatistics}.
}
