% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCAnalysis.R
\name{getNCAnalysis}
\alias{getNCAnalysis}
\title{API wrapper for NCA analysis pharmacodynamic data}
\usage{
getNCAnalysis(conc = NULL, time = NULL, dose = NULL, duration = NULL,
  lamznpt = NULL, usepoints = NULL, excpoints = NULL,
  inter = c("Linear", "Lin up Log down", "Linear Log"), maxdiffrsq = 1e-04,
  minr2adj = 0.5, numhalflife = 1, type = c("Single", "Multiple"),
  tau = NULL, addt0 = FALSE)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{duration}{Single numeric value of duration of infusion}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}

\item{usepoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to use for calculation of terminal phase.
Used rows are flagged by usepoints as \code{TRUE}.}

\item{excpoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to exclude from automatic calculation of terminal phase.
Excluded rows are flagged by excpoints as \code{TRUE}.}

\item{inter}{Single character stating whether the interpolation method used is
\code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}

\item{maxdiffrsq}{single numeric in range 0-1 The Adjusted R-squared method
will select the number of points
for terminal phase calculation for the set of trailing points with the
most points that is within maxdiffrsq of the maximum adjusted R-squared
(default 1e-4)}

\item{minr2adj}{single numeric Minimum value permitted for
adjusted R-squared for lambda-z to be calculated.
If NA or NULL, lambda-z calculation is supressed. (default 0.5)}

\item{numhalflife}{single numeric Multiplier for terminal phase half life,
when checking that terminal phase half life is not
larger than the range of interpolated data
(i.e. extrapolated area is not a large portion of AUC0_Inf).
If LAMZHL * numhalflife > LAMZUL - LAMZLL, terminal phase
statistics will not be returned.
To always return terminal phase statistics, set to 0.
To never return terminal phase statistics set to Inf, NULL or NA. (default 1)}

\item{type}{single character with value "Single", or "Multiple", does project
contain single dose or multiple dose data? (default "Single")}

\item{tau}{single numeric >0 or NULL must be NULL if type Single,
must be single numeric if type Multiple(default NULL)}

\item{addt0}{single logical should t=0, c=0
be added to time if missing? (default FALSE)}
}
\value{
Data frame
}
\description{
Perform Non-Compartmental Analysis for a concentration-time repeated measures dataset.
}
\details{
This is a wrapper function designed to be the interface for performing a complete Non-Compartmental Analysis.
The function has required arguments: \enumerate{
     \item conc a vector of concentration values of length equal to \code{length(time)}
     \item time a corresponding vector of measurement time values.
     \item dose a single numeric value declaring dose
     \item duration (duration of infusion), a single numeric value declaring duration of infusion.
}
Optional arguments are: \enumerate{
     \item lamznpt, a single integer value declaring number of points to use in \code{lambdazStatistics}
     when calculating terminal phase elimination.
     If lamznpt is 0 (zero), terminal phase calculations are supressed.
     By default, lamznpt is NULL, declaring that lamznpt should automatically be calculated by \code{selectPoints} using the default method.
     \item usepoints suppresses automatic selection of terminal phase;
     all rows which are \code{TRUE} will be used to calculate lambdaz. This argument must not be used with lamznpt.
     \item excpoints uses automatic selection, but all rows which are \code{TRUE} are excluded
     from the terminal phase calculation. If usepoints and excpoints are both supplied, then automatic selection will be supressed,
     but excpoints will be omitted in addition to rows excluded by usepoints.
     If supplied with lamznpt, excluded values will not be counted in points used for lambdaz.
     \item inter, a single character stating whether the interpolation method used is \code{"Linear"} (default) or \code{"Linear Log"}
}
No unit conversion is performed by these functions.
\code{getNCAnalysis} returns a data frame with 39 columns containing 1 row of numeric results.
The final column of the return is \code{ERROR}. This should be zero. If it is not, an error
has occured. The text of each error will appear in this column as a single character string with
individual messages separated by a newline.
If there are three or fewer rows, data is passed to \code{\link{ncaPeakTrough}} for Peak/Trough identification only.
If there are four or more rows, data is passed to \code{\link{ncaComplete}} for full non-compartmental analysis.
Returns a data frame with columns:
\enumerate{
     \item \code{R2ADJ}
     \item \code{INTERCEPT}
     \item \code{LAMZNPT}
     \item \code{R2}
     \item \code{CORRXY}
     \item \code{AUCPEO}
     \item \code{AUCPEP}
     \item \code{AUCIFO}
     \item \code{AUCIFP}
     \item \code{AUCLST}
     \item \code{AUMCPEO}
     \item \code{AUMCPEP}
     \item \code{AUMCIFO}
     \item \code{AUMCIFP}
     \item \code{AUMCLST}
     \item \code{CLST}
     \item \code{CLO}
     \item \code{CLP}
     \item \code{CMAX}
     \item \code{CMIN}
     \item \code{INTDOSE}
     \item \code{DOSE}
     \item \code{LAMZHL}
     \item \code{LAMZ}
     \item \code{LAMZLL}
     \item \code{LAMZUL}
     \item \code{MRTIFO}
     \item \code{MRTIFP}
     \item \code{MRTLST}
     \item \code{CPEAK}
     \item \code{TLST}
     \item \code{TMAX}
     \item \code{TMIN}
     \item \code{CTROUGH}
     \item \code{VSSO}
     \item \code{VSSP}
     \item \code{VZO}
     \item \code{VZP}
     \item \code{ERROR}
 }
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
getNCAnalysis(conc = Theoph1$conc, time = Theoph1$Time,
    dose = 10, duration = 1)
}
\author{
Mango Solutions
}
\keyword{nca}
