% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartialAUC.R
\name{getPartialAUC}
\alias{getPartialAUC}
\title{Partial Area Under Curve}
\usage{
getPartialAUC(Conc = NULL, Time = NULL, startTime = NULL,
  endTime = NULL, numPoints = NULL, usePoints = NULL, excPoints = NULL,
  Safe = TRUE, inter = "Linear")
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{startTime}{Single numeric value, first time from which area should be calculated}

\item{endTime}{Single numeric value, last time to which area should be calculated}

\item{numPoints}{single numeric value declaring number of points to use}

\item{usePoints}{If \code{NULL} (default) automatically select, else, logical vector of points to use for calculation of terminal phase. Used rows are flagged by usePoints as \code{TRUE}.}

\item{excPoints}{If \code{NULL} (default) automatically select, else, logical vector of points to exclude from automatic calculation of terminal phase. Excluded rows are flagged by excPoints as \code{TRUE}.}

\item{Safe}{Single logical value, if FALSE don't perform additional error checks}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}
}
\value{
Data frame
}
\description{
Area Under the Concentration Time Curve from T = startTime to T = endTime
}
\details{
This is a wrapper function designed to be the interface for performing a partial AUC calculation. 
\code{getPartialAUC} returns the area under a time-concentration curve from \code{firstTime} up until \code{endTime}. 
The calculation is performed by subtracting AUC_0_startTime from AUC_0_endTime using \code{AUCPartial}.
\code{endTime} need not be one of the elements of \code{Time}.
If \code{endTime} does not coincide with an existing time element and is less than tlast, 
the following interpolation formula will be used to calculated a new concentration: 
    \deqn{c_{inter} = c_1 +  \left| \frac{ t^{end} - t_1}{t_2 - t_1} \right|  (c_2 - c_1)}.  
If endTime is greater than tlast, then another concentration is extrapolated via the formula: 
    \deqn{c_{extr} = c_0 + \exp(- \lambda_z * endTime)  }.
\code{numPoints} is calculated by \code{\link{selectPoints}}.
Returns a data frame with 1 row containing columns:
 \enumerate{
     \item \code{startTime}  
     \item \code{endTime} 
     \item \code{AUCPartial}
     \item \code{numPoints} 
     \item \code{Error} 
 }
}
\author{
Mango Solutions
}
\keyword{math}
