% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCPartial.R
\name{getTerminalAUC}
\alias{getTerminalAUC}
\title{Area Under Terminal Concentration Time Curve}
\usage{
getTerminalAUC(Conc, Time, endTime, lambdaZStats = NULL, numPoints = NULL,
  usePoints = NULL, excPoints = FALSE, minPoints = 3, useObs = FALSE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{endTime}{: single numeric, last time until which area should be calculated}

\item{lambdaZStats}{if not \code{NULL}, a list with 9 elements (see details). Must be provided if numPoints is not.}

\item{numPoints}{: single numeric, number of points to use for lambdaz calculation}

\item{usePoints}{If NULL (default) automatically select, else, logical vector of points to use (\code{TRUE}) for calculation of terminal phase.}

\item{excPoints}{Logical vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}

\item{useObs}{: single logical, use observed rather than predicted CLast, defaults to FALSE}
}
\value{
Single numeric with partial area under curve from TLast to endTime
}
\description{
Area Under Terminal Concentration Time Curve from T = Last to T = endTime
}
\details{
This function calculates the area under the terminal phase exponential
from TLast to endTime. This calculation is the integral of the Terminal 
phase exponential decay function: 
\deqn{\int intercept \ast exp\left ( -\lambda _{z} \ast T\right ) dT = 
    -intercept / \left ( \lambda _{z} \ast exp\left ( \lambda _{z} \ast T\right )\right)}
The partial area under the terminal concentration time curve can be calculated 
by subtracting modeled AUC0Tlast from modeled AUC0endTime.
This is intended to be used as an internal function from \code{\link{AUCPartial}}, so no 
data checking is performed so clean data is assumed (see \code{\link{cleanConcTime}}).
}
\seealso{
\code{\link{lambdaZStatistics}} for more about the numPoints parameter
}
\author{
Mango Business Solutions
}
\keyword{math}
