% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaInitialized.R
\name{getWhichVersion}
\alias{getWhichVersion}
\title{Function to return R version to R.NET}
\usage{
getWhichVersion()
}
\value{
data frame with 1 column, Version and 1 character row
}
\description{
Get R Version
}
\details{
This is a trivial test used to confirm that R is running correctly.
}
\examples{
getWhichVersion()
}
\author{
Mango Solutions
}
\keyword{math}
\keyword{nca}
