% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{isErrorWithMessage}
\alias{isErrorWithMessage}
\title{Check if object is error with given error message}
\usage{
isErrorWithMessage(x, expectedMessage)
}
\arguments{
\item{x}{Object to check}

\item{expectedMessage}{String with expected message}
}
\value{
TRUE or FALSE depending on whether or not x is as expected
}
\description{
Check if an object is of class "try-error", and that it contains a specific error message.
}
\author{
Mango Solutions
}
