% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRTInfSD.R
\name{MRTInfObsSD}
\alias{MRTInfObsSD}
\title{Calculate Mean Residence Time to infinity (observed) for steady-state data, model M3}
\usage{
MRTInfObsSD(Conc, Time, numPoints, Dof = 0, addT0 = FALSE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times.  Must be sorted, and without duplicates}

\item{numPoints}{Number of points to use for lambda z calculation. See \code{\link{lambdaZStatistics}}}

\item{Dof}{Duration of infusion. Must be a}
}
\value{
A single numeric vector.  The value returned is:
\deqn{\frac{AUMCInfObs}{AUCInfObs} - \frac{Dof}{2}}
}
\description{
Mean Residence Time (Single Dose) Observed
}
\details{
Calculates the mean residence time to infinity (observed), for steady-state models of type "M3".
}
\author{
Mango Solutions
}
\keyword{math}
