% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRTInfSD.R
\name{MRTInfPredSD}
\alias{MRTInfPredSD}
\title{Calculate Mean Residence Time to infinity (predicted) for steady-state data, model M3}
\usage{
MRTInfPredSD(Conc, Time, numPoints, Dof = 0, addT0 = FALSE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times.  Must be sorted, and without duplicates}

\item{numPoints}{Number of points to use for lambda z calculation. See \code{\link{lambdaZStatistics}}}

\item{Dof}{Duration of infusion, single numeric value}

\item{addT0}{Single logical value determining whether T = 0 should be added}
}
\value{
A single numeric vector.  The value returned is:
\deqn{\frac{AUMCInfPred}{AUCInfPred} - \frac{Dof}{2}}
}
\description{
Mean Residence Time (Single Dose) Predicted
}
\details{
Calculates the mean residence time to infinity (predicted), for steady-state models of type "M3".
}
\author{
Mango Solutions
}
\keyword{math}
