% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaComplete.R
\name{ncaComplete}
\alias{ncaComplete}
\title{Non-Compartmental Analysis on concentration-time Data}
\usage{
ncaComplete(conc, time, dose, duration, lamznpt = NA_real_,
  usepoints = NULL, excpoints = FALSE, addt0 = FALSE, inter = "Linear",
  ROutput = shapeROutput, maxdiffrsq = 1e-04, minr2adj = 0.8,
  numhalflife = 1)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{duration}{Single numeric value of duration of infusion}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}

\item{usepoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to use for calculation of terminal phase.
Used rows are flagged by usepoints as \code{TRUE}.}

\item{excpoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to exclude from automatic calculation of terminal phase.
Excluded rows are flagged by excpoints as \code{TRUE}.}

\item{addt0}{single logical should t=0, c=0
be added to time if missing? (default FALSE)}

\item{inter}{Single character stating whether the interpolation method used is
\code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}

\item{ROutput}{Named vector consisting of 38 NA and 1 zero
with names which should match expected output names.}

\item{maxdiffrsq}{single numeric in range 0-1 The Adjusted R-squared method
will select the number of points
for terminal phase calculation for the set of trailing points with the
most points that is within maxdiffrsq of the maximum adjusted R-squared
(default 1e-4)}

\item{minr2adj}{single numeric Minimum value permitted for
adjusted R-squared for lambda-z to be calculated.
If NA or NULL, lambda-z calculation is supressed. (default 0.5)}

\item{numhalflife}{single numeric Multiplier for terminal phase half life,
when checking that terminal phase half life is not
larger than the range of interpolated data
(i.e. extrapolated area is not a large portion of AUC0_Inf).
If LAMZHL * numhalflife > LAMZUL - LAMZLL, terminal phase
statistics will not be returned.
To always return terminal phase statistics, set to 0.
To never return terminal phase statistics set to Inf, NULL or NA. (default 1)}
}
\value{
Data frame
}
\description{
Perform Non-Compartmental Analysis for Complete concentration-time Datasets
}
\details{
This function is not expected to be used directly. The output structure
is passed from \code{getNCAnalysis}.
\code{ncaComplete} calculates NCA parameters for individuals returning
as many parameters as are calculable.
\code{CPEAK} and \code{CTROUGH} will always be returned where possible.
The final column of the return is \code{ERROR}. This should be zero.
\code{ncaComplete} will \code{try} to call the functions to calculate the
NCA parameters.
If it is not, an error has occured.
The text of each error will appear in this column as a single character string
with individual messages separated by a newline.
Returns a data frame with 1 row containing columns:
 \enumerate{
     \item \code{R2ADJ}
     \item \code{INTERCEPT}
     \item \code{LAMZNPT}
     \item \code{R2}
     \item \code{CORRXY}
     \item \code{AUCPEO}
     \item \code{AUCPEP}
     \item \code{AUCIFO}
     \item \code{AUCIFP}
     \item \code{AUCLST}
     \item \code{AUMCPEO}
     \item \code{AUMCPEP}
     \item \code{AUMCIFO}
     \item \code{AUMCIFP}
     \item \code{AUMCLST}
     \item \code{CLST}
     \item \code{CLO}
     \item \code{CLP}
     \item \code{CMAX}
     \item \code{CMIN}
     \item \code{INTDOSE}
     \item \code{DOSE}
     \item \code{LAMZHL}
     \item \code{LAMZ}
     \item \code{LAMZLL}
     \item \code{LAMZUL}
     \item \code{MRTIFO}
     \item \code{MRTIFP}
     \item \code{MRTLST}
     \item \code{CPEAK}
     \item \code{TLST}
     \item \code{TMAX}
     \item \code{TMIN}
     \item \code{CTROUGH}
     \item \code{VSSO}
     \item \code{VSSP}
     \item \code{VZO}
     \item \code{VZP}
     \item \code{ERROR}
 }
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
ncaComplete(conc = Theoph1$conc, time = Theoph1$Time,
    dose = Theoph1$Dose[1], duration = 1)
}
\author{
Mango Solutions
}
