% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaInitialized.R
\name{ncaInitialized}
\alias{ncaInitialized}
\title{Function to return R version to R.NET}
\usage{
ncaInitialized(rVersionMinor = 14.1)
}
\arguments{
\item{rVersionMinor}{single numeric or character string describing R version to test against}
}
\value{
data frame
}
\description{
Check mangoNCA is Initialized
}
\details{
This is a second test used to confirm that R is running correctly.
Returns a data frame with 3 columns and 1 row:
 \enumerate{
     \item \code{rVersionOk} logical
     \item \code{rArchOk} logical
     \item \code{statsOk} logical
 }
}
\examples{
ncaInitialized()
}
\author{
Mango Solutions
}
\keyword{math}
\keyword{nca}
