% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaPeaktrough.R
\name{ncaPeakTrough}
\alias{ncaPeakTrough}
\title{PeakTrough Vector of pre- and post-dose identifiers (TROUGHCODE and PEAKCODE respectively). There must not be more than one of each. NULL is replaced with 0.}
\usage{
ncaPeakTrough(Conc, Time, Dose, Dof, PeakTrough = NULL, Safe = TRUE,
  ROutput)
}
\arguments{
\item{Conc}{Vector of concentrations.}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates.}

\item{Dose}{Single numeric value of dose.}

\item{Dof}{Single numeric value of duration of infusion.}

\item{Safe}{Single logical value declaring whether to perform redundant data checks (default is TRUE).}

\item{ROutput}{Named vector consisting of 38 NA and 1 zero with names which should match expected output names.}
}
\value{
Data frame
}
\description{
Return Concentration values coded as PEAKCODE or TROUGHCODE for sparse data
}
\details{
Calculates Peak through (pre / post dose concentrations) for individuals on Time concentration curves.
PEAKCODE is 2, TROUGHCODE is 1, if there is more than one of each of these, this is a data error.
}
\examples{
load(system.file(package = "MangoNca", "data", "shapeROutput.RData"))
Theoph1 <- subset(Theoph, Subject == 1)
ncaPeakTrough(Conc = Theoph1$conc, Time = Theoph1$Time, Dose = Theoph1$Dose[1], Dof = 1, PeakTrough = c(1, 0, 0, 0, 2, rep(0, 6)), ROutput = shapeROutput)
}
\author{
Mango Solutions
}
