% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcExtrap.R
\name{pcExtrap}
\alias{pcExtrap}
\title{Percentage Extrapolated}
\usage{
pcExtrap(Whole, Measured)
}
\arguments{
\item{Whole}{Value of whole}

\item{Measured}{Value of Whole which was measured}
}
\value{
Single numeric value (percentage) of Whole which was extrapolated
}
\description{
Percentage of Whole Extrapolated.
}
\note{
The formula used to calculate this 
quantity is \deqn{ 100 * (Whole - Measured) / Whole }
}
\examples{
pcExtrap(Whole = 100, Measured = 33)
}
\author{
Mango Solutions
}
\keyword{math}
