% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{runMangoNcaTests}
\alias{runMangoNcaTests}
\title{Run MangoNca Test Suite}
\usage{
runMangoNcaTests(testPath = system.file(package = "MangoNca", "testing"))
}
\arguments{
\item{testPath}{A single string holding the path to the internal test suite scripts}
}
\value{
suiteResults
}
\description{
Execute the RUnit unit tests suites for the MangoNca package.  
Prints an html report with name "Unit_Test_Report_MangoNca_<maj>_<min>_<svn>.html"
The package version is extracted from the DESCRIPTION file
}
\examples{
\dontrun{ 
     runMangoNcaTests()
}
}
\author{
Mango Solutions
}
\keyword{debugging}
\keyword{programming}
