% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaComplete.R
\docType{data}
\name{shapeROutput}
\alias{shapeROutput}
\title{Expected Output Shape}
\format{data frame with columns:
 \itemize{
     \item \code{R2ADJ}
     \item \code{INTERCEPT}
     \item \code{LAMZNPT}
     \item \code{R2}
     \item \code{CORRXY}
     \item \code{AUCPEO}
     \item \code{AUCPEP}
     \item \code{AUCIFO}
     \item \code{AUCIFP}
     \item \code{AUCLST}
     \item \code{AUMCPEO}
     \item \code{AUMCPEP}
     \item \code{AUMCIFO}
     \item \code{AUMCIFP}
     \item \code{AUMCLST}
     \item \code{CLST}
     \item \code{CLO}
     \item \code{CLP}
     \item \code{CMAX}
     \item \code{CMIN}
     \item \code{INTDOSE}
     \item \code{DOSE}
     \item \code{LAMZHL}
     \item \code{LAMZ}
     \item \code{LAMZLL}
     \item \code{LAMZUL}
     \item \code{MRTIFO}
     \item \code{MRTIFP}
     \item \code{MRTLST}
     \item \code{CPEAK}
     \item \code{TLST}
     \item \code{TMAX}
     \item \code{TMIN}
     \item \code{CTROUGH}
     \item \code{VSSO}
     \item \code{VSSP}
     \item \code{VZO}
     \item \code{VZP}
     \item \code{ERROR}
 }}
\description{
list with names
}
\keyword{datasets}
