% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPoints.R
\name{usedPoints}
\alias{usedPoints}
\title{Use Selection to Calculate Terminal Phase Statistics

Value is NA if selection would be less than minpoints.
The number of points will not usually be related to the index of the
points used, since any rows may be selected by the user}
\usage{
usedPoints(conc, time, usepoints, excpoints = FALSE, minpoints = 3)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{usepoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to use for calculation of terminal phase.
Used rows are flagged by usepoints as \code{TRUE}.}

\item{excpoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to exclude from automatic calculation of terminal phase.
Excluded rows are flagged by excpoints as \code{TRUE}.}

\item{minpoints}{Single numeric declaring minimum number of points
required for lambdaz calculation.}
}
\value{
Single numeric value
}
\examples{
mangoNCA:::usedPoints(conc = c(435, 363, 113, 47,0),
    time = 0:4, usepoints = c(TRUE, TRUE, TRUE, FALSE, FALSE))
mangoNCA:::usedPoints(conc = c(435, 363, 113, 47,0),
    time = 0:4, usepoints = rep(TRUE, 5),
    excpoints = c(FALSE, FALSE, TRUE, FALSE, FALSE))
}
\author{
Mango Solutions
}
