% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSS.R
\name{VSS}
\alias{VSS}
\title{Volume Distribution at Steady-State (observed)}
\usage{
VSS(MRT, CL)
}
\arguments{
\item{MRT}{Single numeric value of time}

\item{CL}{Single numeric value of clearance}
}
\value{
Single numeric value
}
\description{
Calculate Volume of Distribution at Steady-State.
}
\details{
This function simply calculates VSS from MRT and CL when they have already been  calculated.
\deqn{ VSS = MRT * CL }
}
\examples{
VSS(MRT = 20, CL = 0.1)
}
\author{
Mango Solutions
}
\keyword{math}
