% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSS.R
\name{VSSObs}
\alias{VSSObs}
\title{Volume Distribution at Steady-State (observed)}
\usage{
VSSObs(Conc, Time, Dose, Dof, numPoints = NULL, Safe = TRUE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{Dose}{Single numeric with dose amount}

\item{Dof}{duration of infusion}

\item{numPoints}{Number of points to use for the lambda z calculation(s)}
}
\value{
Single numeric value
}
\description{
Function to Calculate Volume of Distribution at Steady State (Observed)
}
\details{
Calculates volume distribution at steady-state (observed).
This is a convenience function for when using the R package. 
\code{\link{ncaComplete}} calls VSS directly since MRTInfObs and CLObs have already been calculated.
\code{VSSObs} calls \code{\link{VSS}} after calling \code{\link{MRTInfObsSD}}, \code{\link{AUCinfobs}} and \code{\link{clearance}}.
}
\note{
All input checks / error handling for CLPred and MRTInfPredSD apply.  The formula used to calculate this 
quantity is \deqn{ MRTPredSD_{\inf} * CLPred }
}
\examples{
Theoph1 <- subset(  Theoph, Subject == 1)
VSSObs( Conc = Theoph1$conc, Time = Theoph1$Time, numPoints = 3, Dose = Theoph1$Dose[1], Dof = 3 )
}
\author{
Mango Solutions
}
\keyword{math}
