% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSS.R
\name{VSSPred}
\alias{VSSPred}
\title{Volume Distribution at Steady-State (predicted)}
\usage{
VSSPred(conc, time, dose, duration, lamznpt = NULL)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{duration}{Single numeric value of duration of infusion}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}
}
\value{
Single numeric value
}
\description{
Function to Calculate Volume of Distribution at Steady State (Predicted).

Calculates volume distribution at steady-state (predicted).
This is a convenience function for when using the R package. 
\code{\link{ncaComplete}} calls VSS since MRTInfPred and CLPred have already been calculated.
\code{VSSPred} calls \code{\link{VSS}} after calling \code{\link{MRTInfPredSD}}, 
\code{\link{AUCInfPred}} and \code{\link{clearance}}.
}
\note{
All input checks / error handling for CLPred and MRTInfPredSD apply.   
The formula used to calculate this quantity is 
\deqn{ MRTPredSD_{\inf} * CLPred }
}
\examples{
Theoph1 <- subset(  Theoph, Subject == 1)
VSSPred(conc = Theoph1$conc, time = Theoph1$time, dose = Theoph1$Dose[1], duration = 3)
}
\author{
Mango Solutions
}
\keyword{math}
