% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VZ.R
\name{VZ}
\alias{VZ}
\title{Volume Distribution at Steady-State (observed)}
\usage{
VZ(lambdaz, AUCInf, dose)
}
\arguments{
\item{MRT}{Single numeric value of time}

\item{CL}{Single numeric value of clearance}
}
\value{
Single numeric value
}
\description{
Calculate Volume of Distribution for single dose from terminal phase.
}
\details{
This function simply calculates VSS from MRT and CL when they have already been  calculated.
\deqn{ VSS = MRT * CL }
}
\examples{
VZ(lambdaz = 0.001, AUCInf = 1000, dose = 10)
}
\author{
Mango Solutions
}
\keyword{math}
