% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VZ.R
\name{VZPred}
\alias{VZPred}
\title{Terminal Volume Distribution, Single Dose (predicted)}
\usage{
VZPred(Conc, Time, Dose, numPoints = NULL)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{Dose}{Single numeric with dose amount}

\item{numPoints}{Number of points to use for the lambda z calculation(s)}

\item{Dof}{duration of infusion (default NULL)}
}
\value{
Single numeric value
}
\description{
Terminal Volume Distribution, Single Dose (predicted)
}
\details{
Calculates volume distribution for single dose from terminal phase (predicted).
}
\note{
All input checks / error handling for LambdaZStatistics and AUCInfPred apply.  
The formula used to calculate this quantity is 
\deqn{ Dose / {\lambda}_z * AUC({\inf})_Pred }
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
VZPred(Conc = Theoph1$conc, Time = Theoph1$Time, Dose = Theoph1$Dose[1])
}
\author{
Mango Solutions
}
\keyword{math}
