% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCLast.R
\name{AUCLast}
\alias{AUCLast}
\title{Calculate Area Under Curve Until Last Measurable Concentration}
\usage{
AUCLast(Conc, Time, addT0 = FALSE, Safe = TRUE, inter = "Linear")
}
\arguments{
\item{Conc}{A numeric vector of concentration values}

\item{Time}{A numeric vector of time values, parallel to \code{Conc} and of the same length.  These should be sorted in ascending order.}

\item{addT0}{Single logical value declaring whether T0 should be added if missing provided execution is Safe (default TRUE).}

\item{Safe}{Single logical value declaring whether to perform data checks and data cleaning (default is TRUE).}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default), 
\code{"Lin up Log down"} or \code{"Linear Log"} (not case or space sensitive).}
}
\value{
single numeric - area under the concentration/time curve until the last measurable concentration.
}
\description{
Area Under the Curve to from Time = 0 to Time = Last
}
\details{
AUCLast calculates the area under a time/concentration curve from the first point up until the last 
measurable concentration using one of three methods.  
\code{Linear} rule calculates the area of the trapezia bounded by each sequential pair of time and concentration points (see \code{AUCLin}).

\code{Linear Log} rule is a combination of the Linear trapezoidal rule and the Log trapezoidal rule (see \code{AUCLog}). Before Cmax  
(as in the absorption phase), the linear trapezoidal method is used. After Cmax  
(as in the elimination phase), the logarithmic trapezoidal method is used.

\code{Lin Up Log Down} rule is a combination of the Linear trapezoidal rule and the Log trapezoidal rule. When concentrations are increasing  
(as in the absorption phase), the linear trapezoidal method is used. When concentrations are decreasing  
(as in the elimination phase), the logarithmic trapezoidal method is used.

These alternative methods to the Linear  is often preferred because  
the linear method is often a good approximation of drug absorption while logarithmic decline is well represented  
by the logarithmic trapezoidal method during drug elimination.

Calculations are performed for the area bounded by T0 and the time of the last nonzero, non-missing concentration.
For comparison with WinNonLin, a timepoint at T = 0 is added if it does not exist. 
Missing values are removed before further calculations are performed.
}
\note{
The following error checks / processing will be performed:
 \enumerate{
     \item If \code{Conc} or \code{Time} are of length 1 or 0, a single NA is returned
     \item \code{Conc} and \code{Time} should be equal length numeric vectors, otherwise an exception will be generated.
     \item If \code{clast} or \code{tlast} is NA, NA will be returned
     \item There should be no duplicated time entries
 }
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
AUCLast(Theoph1$conc, Theoph1$Time)
}
\author{
Mango Solutions
}
\keyword{math}
