% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUCLog}
\alias{AUCLog}
\title{Calculate AUC Using Logarithmic Trapezoidal Rule}
\usage{
AUCLog(conc, time)
}
\arguments{
\item{conc}{numeric vector of concentrations}

\item{time}{numeric vector of time points (same length as and parallel to conc).
These elements should be sorted in ascending order, otherwise an exception will be generated.}
}
\value{
numeric vector of trapezium AUC values.
}
\description{
Calculates the area under a time-concentration curve (AUC) using the logarithmic trapezoidal rule.
}
\details{
The logarithmic trapezoidal method uses logarithmic interpolation between data points to calculate the AUC. 
This method represents more closely decreasing concentrations since they are equivalent to first order  
elimination  (linear on a logarithmic scale). This function is called by \code{AUCLast}.

A vector of areas under individual trapezoids (not the total area) is returned.
These will be the individual AUC values at a given time/concentration index.
In other words, auc[i] is the area under the trapezium bounded by time[i], conc[i] and time[i + 1], conc[i + 1]. 
For a given time interval (t1:t2), the AUC can be calculated as follows:
    AUC_t1_t2 = (t2 - t1) * (C2 - C1) / log(C2 / C1)
}
\note{
It is assumed that these times and concentrations come from a single curve. 
The following error checks / processing will be performed:
 \enumerate{
     \item If \code{conc} or \code{time} are of length 1 or 0, a single NA is returned
     \item \code{conc} and \code{time} should be equal length numeric vectors, 
     otherwise an exception will be generated
     \item if \code{time[i]} or \code{conc[i]} is \code{NA} and \code{auc =  AUCLin(conc,time)}, 
     then \code{auc[i]} and \code{auc[i - 1]} will be \code{NA}.
     \item If \code{conc[i]} equals \code{conc[i + 1]}, linear interpolation will be used
 }
}
\examples{
Theoph2 <- subset(Theoph, Subject == 2)
x <- mangoNCA:::AUCLog( Theoph2$conc, Theoph2$time ) 
sum(x) # 
}
\keyword{math}
