% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCPartial.R
\name{AUCPartial}
\alias{AUCPartial}
\title{Partial Area Under concentration time Curve}
\usage{
AUCPartial(conc, time, endtime, lamznpt = NULL, lambdaZStats = NULL,
  usepoints = NULL, excpoints = FALSE, minpoints = 3, addt0 = FALSE,
  inter = "Linear", useObs = FALSE, maxdiffrsq = 1e-04, minr2adj = 0.8,
  numhalflife = 1)
}
\arguments{
\item{conc}{Vector of concentrations}

\item{time}{Vector of times, must be ordered in ascending order and 
should not have duplicates}

\item{endtime}{: last time at which area should be calculated}

\item{lamznpt}{: Number of points to use for lambda-z calculation.}

\item{lambdaZStats}{if not \code{NULL}, a list with 9 elements (see details). 
Must be provided if lamznpt is not.}

\item{usepoints}{If NULL (default) automatically select, else, logical 
vector of points to use (\code{TRUE}) for calculation of terminal phase.}

\item{excpoints}{Logical vector (\code{FALSE} by default) excludeing 
\code{TRUE} rows from automatic calculation of terminal phase.}

\item{minpoints}{single numeric, the fewest points for which a linear model 
fit should be attempted (default 3)}

\item{addt0}{single logical, should T=0, C=0 be added to the input data if 
T0 is missing? (default \code{FALSE})}

\item{inter}{Single character stating whether the interpolation method used 
is \code{"Linear"} (default), \code{"Lin up Log down"} or \code{"Linear Log"}}

\item{useObs}{single logical, use observed rather than predicted CLast, 
(default FALSE)}

\item{maxdiffrsq}{single numeric in range 0-1 The Adjusted R-squared method 
will select the number of points
for terminal phase calculation for the set of trailing points with the 
most points that is within maxdiffrsq of the maximum adjusted R-squared 
(default 1e-4)}

\item{minr2adj}{single numeric Minimum value permitted for adjusted 
R-squared for lambda-z to be calculated. 
If NA or NULL, lambda-z calculation is supressed. (default 0.8)}

\item{numhalflife}{single numeric Multiplier for terminal phase half life, 
when checking that terminal phase half life is not an excessively 
large portion of the AUC0_Inf. 
To always return lambda-z, set to 0.
To never return lambda-z set to Inf, NULL or NA. (default 1)}
}
\value{
Single numeric with partial area under curve
}
\description{
Area Under the concentration time Curve from T = 0 to T = endtime
}
\details{
Calculates the area under a time-concentration curve from the first time up until the \code{endtime}.  
\code{endtime} need not be one of the elements of \code{time}, but it should not lie before the minimum time.
If the endtime does not coincide with an existing time element and is less than tlast, the following 
interpolation formula will be used to calculated a new concentration: 
\deqn{c_{inter} = c_1 +  \left| \frac{ t^{end} - t_1}{t_2 - t_1} \right|  (c_2 - c_1)}.  
If endtime is greater than tlast the extrapolated area is calculated using \code{getTerminalAUC} 
which calculates the area under the terminal phase exponential from TLast to endtime.
\code{lamznpt} is calculated by \code{\link{getPartialAUC}} .  
This calculation is the integral of the Terminal phase exponential decay function: 
\deqn{\int intercept \ast exp\left ( -\lambda _{z} \ast T\right ) dT = 
    -intercept / \left ( \lambda _{z} \ast exp\left ( \lambda _{z} \ast T\right )\right)}
The partial area under the terminal concentration time curve can be calculated 
by subtracting modeled AUC0Tlast from modeled AUC0endtime.
\code{lamznpt} may be NA provided endtime < Tlast.
}
\seealso{
\code{\link{lambdaZStatistics}} for more about the lamznpt parameter
}
\author{
Mango Business Solutions
}
\keyword{math}
