% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUMCPartial.R
\name{AUMCPartial}
\alias{AUMCPartial}
\title{Partial Area Under Moment Curve}
\usage{
AUMCPartial(Conc, Time, endTime, numPoints, addT0 = TRUE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{endTime}{: last time at which area should be calculated.  Must be greater than the smallest element of Time, else NA is returned.}

\item{numPoints}{: Number of points to use for lambda-z calculation.  This will only be used if endTime > tlast to 
extrapolate a new concentration point}
}
\value{
Single numeric
}
\description{
Area Under the Concentration Time Moment Curve from T = 0 to T = endTime
}
\details{
Calculates the area under a concentration-time moment curve from the first time up until the \code{endTime}.  \code{endTime}
need not be one of the elements of \code{Time}, but it should like between the minimum and maximum times 
Note, all error checks that apply to \code{\link{AUC}} apply here.  If \code{endTime} is less than
\code{tmin} \code{NA} is returned.  There should be no missing times.
If the endTime does not coincide with an existing time element, the following interpolation formula will be
used to calculated a new concentration (if time is less than tlast): 
\deqn{c_{inter} = c_1 +  \left| \frac{ t^{end} - t_1}{t_2 - t_1} \right|  (c_2 - c_1)}.  If endTime is greater than
tlast, then another concentration is extrapolated via the formula \deqn{c_{extr} = c_0 + \exp(- \lambda_z * endTime)  }
}
\author{
Mango Solutions
}
\keyword{math}
