% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaComplete.R
\name{chooseNumPointsAction}
\alias{chooseNumPointsAction}
\title{Choose how to handle numPoints}
\usage{
chooseNumPointsAction(Conc = Conc, Time = Time,
  numPoints = as.numeric(NA), usePoints = NULL, excPoints = FALSE)
}
\arguments{
\item{Conc}{Vector of concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{numPoints}{single numeric value declaring number of points to use}

\item{usePoints}{If \code{NULL} (default) automatically select, else, logical vector of points to use for calculation of terminal phase. Used rows are flagged by usePoints as \code{TRUE}.}

\item{Logical}{vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}
}
\value{
length 2 list with elements ACTION, a single character value, and MINROWSFORLAMBDAZ, a single numeric with value 3
}
\description{
Decide whether to use or automatically calculate numPoints
}
\author{
Mango Solutions
}
\keyword{math}
