% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmintmin.R
\name{CminTmin}
\alias{CminTmin}
\title{Calculate Minimum concentration and time of minimum concentration}
\usage{
CminTmin(conc, time)
}
\arguments{
\item{conc}{numeric vector of concentrations}

\item{time}{numeric vector of time points (parallel to \code{conc})}
}
\value{
list with numeric scalar (length 1 vector) components "cmin", "tmin" and "index".
}
\description{
Finds the minimum concentration value and the first time point (in the sense of corresponding to the minimal index)
with that concentration.
}
\note{
The following error checks / processing will be performed:
 \enumerate{
     \item If \code{conc} or \code{time} are of length 0, NA is returned for both values
     \item If the sum of \code{Code} (after omitting missing values) is 0, NA is returned for both values
     \item \code{cmin} is computed by stripping NA values from \code{conc}
     \item if \code{conc} and \code{time} are not equal length numeric vectors, an exception will be generated by \code{checkNumericSameLength}
 }
}
\examples{
Theoph2 <- subset(Theoph, Subject == 2)
CminTmin( Theoph2$conc, Theoph2$time ) 
}
\keyword{math}
