% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPoints.R
\name{fixedPoints}
\alias{fixedPoints}
\title{Fixed Points for Calculating Terminal Rate Constant}
\usage{
fixedPoints(conc, time, lamznpt, excpoints = FALSE, minpoints = 3)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}

\item{excpoints}{If \code{NULL} (default) automatically select, else,
logical vector of points to exclude from automatic calculation of terminal phase.
Excluded rows are flagged by excpoints as \code{TRUE}.}

\item{minpoints}{Minimum number of points to use for the lambda-z calculation(s),
3 by default. Single positive integer.}
}
\value{
List consisting of lamznpt, a single numeric value, and result,
a length 9 numeric vector.
}
\description{
Use Fixed Number of Trailing Points for Terminal Rate Constant (lambda-z) Calculation.
}
\details{
Returns input lamznpt and results for Terminal Rate Constant (lambda-z) calculation
fixedPoints calls \code{\link{lambdaZStatistics}} to perform regressions
using the specified number of trailing non-zero rows
Rows prior to Cmax may be used.
Exclusion of rows prior to the end of infusion is not currently implemented.
Zero or missing values will error.
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
mangoNCA:::fixedPoints(conc = Theoph1$conc, time = Theoph1$Time, lamznpt = 3)
}
\author{
Mango Solutions
}
