% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaComplete.R
\name{ncaComplete}
\alias{ncaComplete}
\title{Non compartment analysis on Time Concentration data}
\usage{
ncaComplete(Conc, Time, Dose, Dof, PeakTrough = NULL,
  numPoints = as.numeric(NA), usePoints = NULL, excPoints = FALSE,
  Safe = TRUE, addT0 = FALSE, inter = "Linear", ROutput)
}
\arguments{
\item{Conc}{Vector of Concentrations}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates}

\item{Dose}{Single numeric of amount of compound delivered to system}

\item{Dof}{Single numeric of duration of infusion.}

\item{PeakTrough}{Optional vector of numeric coding of length Time (default is NULL).}

\item{numPoints}{If NA (default) number of points for calculation of terminal phase is calculated by \code{selectPoints}, otherwise a single integer declaring number of points to use. If numPoints is zero or less, supress terminal phase calculations}

\item{usePoints}{If NULL (default) automatically select, else, logical vector of points to use (\code{TRUE}) for calculation of terminal phase.}

\item{excPoints}{Logical vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}

\item{Safe}{Single logical value declaring whether to perform redundant data checks (default is TRUE).}

\item{addT0}{Single logical value declaring whether to add row at T = 0 if missing (default is FALSE).}

\item{inter}{Single character stating whether the interpolation method used is \code{"Linear"} (default) or \code{"Linear Log"}}

\item{ROutput}{Named vector consisting of 38 NA and 1 zero with names which should match expected output names.}
}
\value{
Data frame
}
\description{
Perform Non-Compartmental Analysis for Complete Concentration-Time Datasets
}
\details{
This function is not expected to be used directly. The output structure is passed from \code{ncaAnalysis}.
\code{ncaComplete} calculates NCA parameters for individuals returning as many parameters as are calculable.
\code{ROutput_Peak} and \code{ROutput_Trough} will always be returned where possible.
The final column of the return is \code{ROutput_Error}. This should be zero. 
\code{ncaComplete} will \code{try} to call the functions to calculate the NCA parameters.
If it is not, an error has occured. 
The text of each error will appear in this column as a single character string with  
individual messages separated by a newline. 
Returns a data frame with 1 row containing columns:
 \enumerate{
     \item \code{ROutput_adjr2}  
     \item \code{ROutput_intercept} 
     \item \code{ROutput_numPoints}  
     \item \code{ROutput_r2}  
     \item \code{ROutput_rhoXY}  
     \item \code{ROutput_AUC_Percent_Extrapolated_obs}  
     \item \code{ROutput_AUC_Percent_Extrapolated_pred}  
     \item \code{ROutput_AUCInfObs}  
     \item \code{ROutput_AUCInfPred}  
     \item \code{ROutput_AUCLast}  
     \item \code{ROutput_AUMC_Percent_Extrapolated_obs}  
     \item \code{ROutput_AUMC_Percent_Extrapolated_pred} 
     \item \code{ROutput_AUMCInfObs}  
     \item \code{ROutput_AUMCInfPred}  
     \item \code{ROutput_AUMCLast}  
     \item \code{ROutput_CLast}  
     \item \code{ROutput_ClObs}  
     \item \code{ROutput_ClPred}  
     \item \code{ROutput_Cmax}  
     \item \code{ROutput_Cmin}  
     \item \code{ROutput_Dof}  
     \item \code{ROutput_Dose}  
     \item \code{ROutput_HalfLife}  
     \item \code{ROutput_LambdaZ}  
     \item \code{ROutput_LambdazLower}  
     \item \code{ROutput_LambdazUpper}  
     \item \code{ROutput_MRTInfObs}  
     \item \code{ROutput_MRTInfPred}  
     \item \code{ROutput_MRTLast}  
     \item \code{ROutput_Peak}  
     \item \code{ROutput_TLast}  
     \item \code{ROutput_Tmax}  
     \item \code{ROutput_Tmin}  
     \item \code{ROutput_Trough}  
     \item \code{ROutput_VssObs}  
     \item \code{ROutput_VssPred}  
     \item \code{ROutput_VzObs} 
     \item \code{ROutput_VzPred} 
     \item \code{ROutput_Error} 
 }
}
\examples{
load(system.file(package = "MangoNca", "data", "shapeROutput.RData"))
Theoph1 <- subset(Theoph, Subject == 1)
ncaComplete(Conc = Theoph1$conc, Time = Theoph1$Time, Dose = Theoph1$Dose[1], Dof = 1, ROutput = shapeROutput)
}
\author{
Mango Solutions
}
