% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaPeaktrough.R
\name{ncaPeakTrough}
\alias{ncaPeakTrough}
\title{PeakTrough Vector of pre- and post-dose identifiers (TROUGHCODE and PEAKCODE respectively). There must not be more than one of each. NULL is replaced with 0.}
\usage{
ncaPeakTrough(conc, time, dose, duration, output)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{duration}{Single numeric value of duration of infusion}

\item{output}{Named vector consisting of 38 NA and 1 zero with names which should match expected output names.}
}
\value{
Data frame
}
\description{
Return concentration values coded as PEAKCODE or TROUGHCODE for sparse data
}
\details{
Calculates Peak through (pre / post dose concentrations) for individuals on time concentration curves.
PEAKCODE is 2, TROUGHCODE is 1, if there is more than one of each of these, this is a data error.
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
ncaPeakTrough(conc = Theoph1$conc, time = Theoph1$time, 
    dose = Theoph1$Dose[1], duration = 1)
}
\author{
Mango Solutions
}
