% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPoints.R
\name{selectPoints}
\alias{selectPoints}
\title{Number of Points for Calculating Terminal Rate Constant}
\usage{
selectPoints(Conc, Time, minPoints = 3, method = "ars", excPoints = FALSE,
  Safe = TRUE)
}
\arguments{
\item{Conc}{Vector of concentrations.}

\item{Time}{Vector of times, must be ordered in ascending order and should not have duplicates.}

\item{minPoints}{Minimum number of points to use for the Lambdaz calculation(s), 3 by default. Single positive integer.}

\item{method}{Method of selecting points (only Adjusted R-Squared Method implemented).}

\item{excPoints}{Logical vector (\code{FALSE} by default) excludeing \code{TRUE} rows from automatic calculation of terminal phase.}

\item{Safe}{Single logical value declaring whether to perform redundant data checks (default is TRUE).}
}
\value{
List consisting of numPoints, a single numeric value, and result, a length 9 numeric vector.
}
\description{
Choose Number of Trailing Points to use for Terminal Rate Constant (Lambdaz) Calculation.
}
\details{
Calculates number of points to use for Terminal Rate Constant (Lambdaz) calculation based on Adjusted R-squared method.
selectPoints calls \code{\link{lambdaZStatistics}} to perform regressions using the last three rows, 
then the last four rows, last five, etc. Rows prior to Cmax are not used.
Exclusion of rows prior to the end of infusion is not currently implemented.
For each regression, an adjusted R2 is computed. The regression with the largest adjusted R2 is 
selected to estimate lambdaz, with these caveats:
 \enumerate{
     \item Number of rows with the greatest number of rows AND within MAXDIFFRSQ (0.0001) of the largest adjusted R2 value is selected.
     \item Rows where Conc <= 0 are not used in the calculation
     \item Trailing rows where Conc <= 0 are removed by \code{stripTrailingZeros} and not counted in number returned
     \item Cmax is not included in Lambdaz calculation (User Requirements PEX-010)
     \item lambdaz must be greater than MINLAMBDAZ (0)
     \item lambdaz must be calculated using at least minPoints (3) data rows
     \item adjr2 must be at least MINADJR2 (0.8)
     \item tPhaseHalfLife must be less than lambdaz interval (LambdazUpper - LambdazLower)
 }
Zero or missing values will be removed with a warning.
}
\examples{
     Theoph1 <- subset(Theoph, Subject == 1)
     selectPoints(Conc = Theoph1$conc, Time = Theoph1$Time)
}
\author{
Mango Solutions
}
\keyword{math}
