% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSS.R
\name{VSSObs}
\alias{VSSObs}
\title{Function to Calculate Volume of Distribution at Steady State (Observed)}
\usage{
VSSObs(conc, time, dose, duration, lamznpt = NULL)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{duration}{Single numeric value of duration of infusion}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}
}
\value{
Single numeric value
}
\description{
Calculates volume distribution at steady-state (observed).
This is a convenience function for when using the R package. 
\code{\link{ncaComplete}} calls VSS directly since MRTInfObs and CLObs have already been calculated.
\code{VSSObs} calls \code{\link{VSS}} after calling \code{\link{MRTInfObsSD}}, \code{\link{AUCInfObs}} and \code{\link{clearance}}.
}
\note{
All input checks / error handling for CLPred and MRTInfPredSD apply.  The formula used to calculate this 
quantity is \deqn{ MRTPredSD_{\inf} * CLPred }
}
\examples{
Theoph1 <- subset(  Theoph, Subject == 1)
VSSObs(conc = Theoph1$conc, time = Theoph1$time, lamznpt = 3, dose = Theoph1$Dose[1], duration = 3 )
}
\author{
Mango Solutions
}
\keyword{math}
