% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VZ.R
\name{VZPred}
\alias{VZPred}
\title{Terminal Volume Distribution, Single dose (predicted)

Terminal Volume Distribution, Single dose (predicted)

Calculates volume distribution for single dose from terminal phase (predicted).}
\usage{
VZPred(conc, time, dose, lamznpt = NULL)
}
\arguments{
\item{conc}{Vector of conc}

\item{time}{Vector of time, must be ordered in ascending order and should not have duplicates}

\item{dose}{Single numeric value of dose}

\item{lamznpt}{If \code{NULL} (default) automatically select, else,
single numeric number of points for calculation of terminal phase.}
}
\value{
Single numeric value
}
\note{
All input checks / error handling for LambdaZStatistics and AUCInfPred apply.  
The formula used to calculate this quantity is 
\deqn{ dose / {\lambda}_z * AUC({\inf})_Pred }
}
\examples{
Theoph1 <- subset(Theoph, Subject == 1)
VZPred(conc = Theoph1$conc, time = Theoph1$time, dose = Theoph1$Dose[1])
}
\author{
Mango Solutions
}
\keyword{math}
